% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{barray_to_edgelist}
\alias{barray_to_edgelist}
\title{Retrieves the edgelist of a barry_graph}
\usage{
barray_to_edgelist(x)
}
\arguments{
\item{x}{An object of class barry_graph.}
}
\value{
A matrix with two columns, the first one with the source and the
second one with the target.
}
\description{
Retrieves the edgelist of a barry_graph
}
\examples{
data(krackhardt_advice)
data(krackhardt_advice_perceptions)

n_people <- 21
advice_matrix <- matrix(0L, nrow = n_people, ncol = n_people)
advice_matrix[cbind(krackhardt_advice$from, krackhardt_advice$to)] <-
  krackhardt_advice$value

krack_graph <- new_barry_graph(
  c(list(advice_matrix), krackhardt_advice_perceptions)
)
head(barray_to_edgelist(krack_graph))
}
