% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_imaginary_census.R
\name{count_imaginary_census}
\alias{count_imaginary_census}
\title{Computes census of imaginary errors}
\usage{
count_imaginary_census(x, counter_type = 0L)
}
\arguments{
\item{x}{An object of class \link{barry_graph}.}

\item{counter_type}{An integer indicating the type of census to compute (see details).}
}
\value{
A data frame of class \code{"imaginary_census"} with columns \code{id},
\code{name}, and \code{value}.
}
\description{
Computes census of imaginary errors
}
\details{
We can also separate the counts as a function of whether the perceiver is looking
into all ties, only ties including them, or only ties not including them.
This is controlled by the \code{counter_type} argument:
\itemize{
\item 0: All ties
\item 1: Only ties including the perceiver
\item 2: Only ties not including the perceiver
}

There are ten (10) values:
\itemize{
\item (01) Accurate null
\item (02) Partial false positive (null)
\item (03) Complete false positive (null)
\item (04) Partial false negative (assym)
\item (05) Accurate assym
\item (06) Mixed assym
\item (07) Partial false positive (assym)
\item (08) Complete false negative (full)
\item (09) Partial false negative (full)
\item (10) Accurate full
}
}
\examples{
data(krackhardt_advice)
data(krackhardt_advice_perceptions)

n_people <- 21
advice_matrix <- matrix(0L, nrow = n_people, ncol = n_people)
advice_matrix[cbind(krackhardt_advice$from, krackhardt_advice$to)] <-
  krackhardt_advice$value

krack_graph <- new_barry_graph(
  c(list(advice_matrix), krackhardt_advice_perceptions)
)
census <- count_imaginary_census(krack_graph)
head(census)
summary(census)
}
