% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{krackhardt_advice_perceptions}
\alias{krackhardt_advice_perceptions}
\title{Krackhardt Advice Network Perception Errors}
\format{
A list of 21 matrices (21x21 each) representing individual perceptions
}
\source{
Generated perception errors based on Krackhardt advice data
}
\usage{
krackhardt_advice_perceptions
}
\description{
Individual perceptions of the Krackhardt advice network containing
systematic perception errors and biases.
}
\keyword{datasets}
