% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{krackhardt_friendship_perceptions}
\alias{krackhardt_friendship_perceptions}
\title{Krackhardt Friendship Network Perception Errors}
\format{
A list of 21 matrices (21x21 each) representing individual perceptions
}
\source{
Generated perception errors based on Krackhardt friendship data
}
\usage{
krackhardt_friendship_perceptions
}
\description{
Individual perceptions of the Krackhardt friendship network containing
systematic perception errors and biases.
}
\keyword{datasets}
