% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barry_graph.R
\name{new_barry_graph}
\alias{new_barry_graph}
\alias{barry_graph}
\alias{new_barry_graph.matrix}
\alias{new_barry_graph.list}
\title{Binary Array Graph}
\usage{
new_barry_graph(x, ...)

\method{new_barry_graph}{matrix}(x, n, ...)

\method{new_barry_graph}{list}(x, ...)
}
\arguments{
\item{x}{Either a matrix or a list of matrices.}

\item{...}{Currently ignored.}

\item{n}{Integer. The size of the original network.}
}
\description{
Binary Array Graph
}
\details{
When \code{x} is a matrix, it is assumed that it will be a block
diagonal matrix, with the first block corresponding to the reference
(true) network.

If \code{x} is a list, the first matrix is assumed to be the reference
(true) network.
}
\examples{
# Create a simple 3x3 network
net1 <- matrix(c(0, 1, 0,
                 1, 0, 1,
                 0, 1, 0), nrow = 3, byrow = TRUE)

# Create another 3x3 network (someone's perception)
net2 <- matrix(c(0, 1, 1,
                 1, 0, 0,
                 1, 0, 0), nrow = 3, byrow = TRUE)

# Method 1: Using a list of matrices
networks_list <- list(net1, net2)
graph1 <- new_barry_graph(networks_list)
print(graph1)

# Method 2: Using a block diagonal matrix
block_matrix <- rbind(
  cbind(net1, matrix(0, 3, 3)),
  cbind(matrix(0, 3, 3), net2)
)
graph2 <- new_barry_graph(block_matrix, n = 3)
print(graph2)
# Example with block diagonal matrix
# Create two 4x4 networks
true_net <- matrix(c(0, 1, 1, 0,
                     1, 0, 0, 1,
                     1, 0, 0, 1,
                     0, 1, 1, 0), nrow = 4, byrow = TRUE)

perceived_net <- matrix(c(0, 1, 0, 0,
                         1, 0, 1, 1,
                         0, 1, 0, 0,
                         0, 1, 0, 0), nrow = 4, byrow = TRUE)

# Combine into block diagonal matrix
block_diag <- matrix(0, 8, 8)
block_diag[1:4, 1:4] <- true_net
block_diag[5:8, 5:8] <- perceived_net

# Create barry graph
graph <- new_barry_graph(block_diag, n = 4)
print(graph)

# Example with list of matrices
# Create friendship network data
friendship <- matrix(c(0, 1, 1, 0,
                      1, 0, 1, 1,
                      1, 1, 0, 0,
                      0, 1, 0, 0), nrow = 4, byrow = TRUE)

# Person 1's perception of the network
person1_view <- matrix(c(0, 1, 1, 1,
                        1, 0, 1, 0,
                        1, 1, 0, 1,
                        1, 0, 1, 0), nrow = 4, byrow = TRUE)

# Person 2's perception of the network  
person2_view <- matrix(c(0, 1, 0, 0,
                        1, 0, 1, 1,
                        0, 1, 0, 0,
                        0, 1, 0, 0), nrow = 4, byrow = TRUE)

# Create list with true network first, then perceptions
network_data <- list(friendship, person1_view, person2_view)

# Create barry graph
social_graph <- new_barry_graph(network_data)
print(social_graph)

# Using the Krackhardt data
# Load example network data
net1 <- matrix(c(0, 1, 1, 0,
                 1, 0, 0, 1,
                 1, 0, 0, 1,
                 0, 1, 1, 0), nrow = 4, byrow = TRUE)

net2 <- matrix(c(0, 1, 0, 0,
                 1, 0, 1, 1,
                 0, 1, 0, 0,
                 0, 1, 0, 0), nrow = 4, byrow = TRUE)

# Create barry graph with example data
example_networks <- list(net1, net2)
example_graph <- new_barry_graph(example_networks)

}
