% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{barray_to_edgelist}
\alias{barray_to_edgelist}
\title{Retrieves the edgelist of a barry_graph}
\usage{
barray_to_edgelist(x)
}
\arguments{
\item{x}{An object of class barry_graph.}
}
\value{
A matrix with two columns, the first one with the source and the
second one with the target.
}
\description{
Retrieves the edgelist of a barry_graph
}
