% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_imaginary_census}
\alias{count_imaginary_census}
\title{Computes census of imaginary errors}
\usage{
count_imaginary_census(x)
}
\arguments{
\item{x}{An object of class \link{barry_graph}.}
}
\description{
Computes census of imaginary errors
}
\details{
There are ten (10) values:
\itemize{
\item (01) Accurate null
\item (02) Partial false positive (null)
\item (03) Complete false positive (null)
\item (04) Partial false negative (assym)
\item (05) Accurate assym
\item (06) Mixed assym
\item (07) Partial false positive (assym)
\item (08) Complete false negative (full)
\item (09) Partial false negative (full)
\item (10) Accurate full
}
}
