% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_recip_errors}
\alias{count_recip_errors}
\title{Add a counter for reciprocity errors}
\usage{
count_recip_errors(x, counter_type = 0L)
}
\arguments{
\item{x}{An object of class \link{barry_graph}.}
}
\description{
Add a counter for reciprocity errors
}
