% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barry_graph.R
\name{new_barry_graph}
\alias{new_barry_graph}
\alias{new_barry_graph.matrix}
\alias{new_barry_graph.list}
\title{Binary Array Graph}
\usage{
new_barry_graph(x, ...)

\method{new_barry_graph}{matrix}(x, n, ...)

\method{new_barry_graph}{list}(x, ...)
}
\arguments{
\item{x}{Either a matrix or a list of matrices.}

\item{...}{Currently ignored.}

\item{n}{Integer. The size of the original network.}
}
\description{
Binary Array Graph
}
\details{
When \code{x} is a matrix, it is assumed that it will be a block
diagonal matrix, with the first block corresponding to the reference
(true) network.

If \code{x} is a list, the first matrix is assumed to be the reference
(true) network.
}
