% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barry_graph.R
\name{new_barry_graph}
\alias{new_barry_graph}
\alias{barry_graph}
\alias{new_barry_graph.matrix}
\alias{new_barry_graph.list}
\alias{netsize}
\alias{nnets}
\title{Binary Array Graph}
\usage{
new_barry_graph(x, ...)

\method{new_barry_graph}{matrix}(x, n, ...)

\method{new_barry_graph}{list}(x, ...)

netsize(x)

nnets(x)
}
\arguments{
\item{x}{Either a matrix or a list of matrices, or an object of class
\code{barry_graph}.}

\item{...}{Currently ignored.}

\item{n}{Integer. The size of the original network.}
}
\value{
The function \code{netsize()} returns the size of individual networks
(all matching).

\code{nnets()} returns the number of graphs contained in the \code{barry_graph}
object.
}
\description{
Binary Array Graph
}
\details{
When \code{x} is a matrix, it is assumed that it will be a block
diagonal matrix, with the first block corresponding to the reference
(true) network.

If \code{x} is a list, the first matrix is assumed to be the reference
(true) network.
}
\examples{
# Using the Krackhardt advice network
data(krackhardt_advice)
data(krackhardt_advice_perceptions)

# Convert edge-list data frame to adjacency matrix
n_people <- 21
advice_matrix <- matrix(0L, nrow = n_people, ncol = n_people)
advice_matrix[cbind(krackhardt_advice$from, krackhardt_advice$to)] <-
  krackhardt_advice$value

krack_graph <- new_barry_graph(
  c(list(advice_matrix), krackhardt_advice_perceptions)
)
krack_graph

# Network size and number of networks
netsize(krack_graph)
nnets(krack_graph)
}
