% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barry_graph.R
\name{print.barry_graph}
\alias{print.barry_graph}
\title{Print Barry Graph}
\usage{
\method{print}{barry_graph}(x, ...)
}
\arguments{
\item{x}{A barry_graph object.}

\item{...}{Additional arguments passed to print (currently ignored).}
}
\value{
Invisibly returns the input object. Called for its side effect of printing.
}
\description{
Print method for barry_graph objects.
}
