% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{tie_level_accuracy}
\alias{tie_level_accuracy}
\title{Null distribution for Cognitive Imaginary Structures}
\usage{
tie_level_accuracy(graph, which_nets = NULL)
}
\arguments{
\item{graph}{A barry_graph object.}

\item{which_nets}{Integer vector. The networks to sample from.}
}
\description{
Null distribution for Cognitive Imaginary Structures
}
\examples{
\dontrun{
taccuracy <- tie_level_accuracy(graph)
boxplot(taccuracy[,-1])
}
}
